pyproject.toml
[project]
name = "scan-installed-packages"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.13.3",
]
